day 20: account for flipped for adjacent edges
In both cases where this is used (i.e. coming from the top to find the
left-edge, or knowing the left-edge and finding the bottom), the edge in
the correct direction depends on whether or not we were flipped getting
to the current location -- the input data would naturally have its
corresponding edge on the other side if we are.

This seems to reconstruct the example image correctly.