day 20: make edge accessors same under rotation
The reconstruction of the image depends on knowing exactly how to rotate
a tile to make matching edges meet up.  Make all four accessors return
the pixels in the same order as-if they had been rotated to the top --
the top is trivially correct, the right-hand side also iterates from top
to bottom, but the left and bottom sides were backwards.

I also observed by drawing-it-out if the edges exactly match
one-another, then they'll be backwards from one another when placed
adjacent (e.g. rotated to the right of one tile and the left of
another).  Inverting the boolean stored in the HashMap means we are
reliably storing "should we flip one of the tiles in order to traverse
this edge in the connectivity graph".

This at least produces a square image after reconstruction now.